regenarate migration file
