--- web_app/alembic/versions/a21c35198b69_added_user_table.py
@@ -1,34 +1,0 @@
-"""Added user table
-
-Revision ID: a21c35198b69
-Revises: 
-Create Date: 2024-10-09 17:03:57.127373
-
-"""
-from alembic import op
-import sqlalchemy as sa
-
-
-# revision identifiers, used by Alembic.
-revision = 'a21c35198b69'
-down_revision = None
-branch_labels = None
-depends_on = None
-
-
-def upgrade() -> None:
-    # ### commands auto generated by Alembic - please adjust! ###
-    op.create_table('user',
-    sa.Column('id', sa.UUID(), nullable=False),
-    sa.Column('is_contract_deployed', sa.Boolean(), nullable=True),
-    sa.Column('wallet_id', sa.String(), nullable=False, index=True),
-    sa.Column('deployed_transaction_hash', sa.String(), nullable=True),
-    sa.PrimaryKeyConstraint('id')
-    )
-    # ### end Alembic commands ###
-
-
-def downgrade() -> None:
-    # ### commands auto generated by Alembic - please adjust! ###
-    op.drop_table('user')
-    # ### end Alembic commands ###
