+++ web_app/alembic/versions/d71e1e3e800f_create_user_and_position_table.py
@@ -1,0 +1,50 @@
+"""create user and position table'
+
+
+Revision ID: d71e1e3e800f
+Revises: 
+Create Date: 2024-10-12 20:41:29.852081
+
+"""
+from alembic import op
+import sqlalchemy as sa
+
+
+# revision identifiers, used by Alembic.
+revision = 'd71e1e3e800f'
+down_revision = None
+branch_labels = None
+depends_on = None
+
+
+def upgrade() -> None:
+    # ### commands auto generated by Alembic - please adjust! ###
+    op.create_table('user',
+    sa.Column('id', sa.UUID(), nullable=False),
+    sa.Column('is_contract_deployed', sa.Boolean(), nullable=True),
+    sa.Column('wallet_id', sa.String(), nullable=False),
+    sa.Column('deployed_transaction_hash', sa.String(), nullable=True),
+    sa.PrimaryKeyConstraint('id')
+    )
+    op.create_index(op.f('ix_user_wallet_id'), 'user', ['wallet_id'], unique=False)
+    op.create_table('position',
+    sa.Column('id', sa.UUID(), nullable=False),
+    sa.Column('user_id', sa.UUID(), nullable=False),
+    sa.Column('token_symbol', sa.String(), nullable=False),
+    sa.Column('amount', sa.String(), nullable=False),
+    sa.Column('multiplier', sa.Integer(), nullable=False),
+    sa.Column('created_at', sa.DateTime(), nullable=False),
+    sa.ForeignKeyConstraint(['user_id'], ['user.id'], ),
+    sa.PrimaryKeyConstraint('id')
+    )
+    op.create_index(op.f('ix_position_user_id'), 'position', ['user_id'], unique=False)
+    # ### end Alembic commands ###
+
+
+def downgrade() -> None:
+    # ### commands auto generated by Alembic - please adjust! ###
+    op.drop_index(op.f('ix_position_user_id'), table_name='position')
+    op.drop_table('position')
+    op.drop_index(op.f('ix_user_wallet_id'), table_name='user')
+    op.drop_table('user')
+    # ### end Alembic commands ###
