--- web_app/alembic/versions/ea10d19e12ca_add_created_at_to_position.py
@@ -1,40 +1,0 @@
-"""add created_at to position
-
-Revision ID: ea10d19e12ca
-Revises: a21c35198b69
-Create Date: 2024-10-11 12:43:42.733767
-
-"""
-from alembic import op
-import sqlalchemy as sa
-
-
-# revision identifiers, used by Alembic.
-revision = 'ea10d19e12ca'
-down_revision = 'a21c35198b69'
-branch_labels = None
-depends_on = None
-
-
-def upgrade() -> None:
-    # ### commands auto generated by Alembic - please adjust! ###
-    op.add_column('position', sa.Column('token_symbol', sa.String(), nullable=False))
-    op.add_column('position', sa.Column('amount', sa.String(), nullable=False))
-    op.add_column('position', sa.Column('multiplier', sa.Integer(), nullable=False))
-    op.add_column('position', sa.Column('created_at', sa.DateTime(), nullable=False))
-    op.create_index(op.f('ix_position_user_id'), 'position', ['user_id'], unique=False)
-    op.create_foreign_key(None, 'position', 'user', ['user_id'], ['id'])
-    op.drop_column('position', 'token_name')
-    # ### end Alembic commands ###
-
-
-def downgrade() -> None:
-    # ### commands auto generated by Alembic - please adjust! ###
-    op.add_column('position', sa.Column('token_name', sa.VARCHAR(), autoincrement=False, nullable=False))
-    op.drop_constraint(None, 'position', type_='foreignkey')
-    op.drop_index(op.f('ix_position_user_id'), table_name='position')
-    op.drop_column('position', 'created_at')
-    op.drop_column('position', 'multiplier')
-    op.drop_column('position', 'amount')
-    op.drop_column('position', 'token_symbol')
-    # ### end Alembic commands ###
